@@ -1,6 +1,6 @@
 {
   "name": "oscars",
-  "version": "0.0.1",
+  "version": "1.0.0-0",
   "scripts": {
     "ng": "ng",
     "start": "node ./bin/www",
@@ -17,65 +17,72 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.2.2",
-    "@angular/common": "10.2.2",
-    "@angular/compiler": "10.2.2",
-    "@angular/core": "10.2.2",
-    "@angular/forms": "10.2.2",
-    "@angular/localize": "10.2.2",
-    "@angular/platform-browser": "10.2.2",
-    "@angular/platform-browser-dynamic": "10.2.2",
-    "@angular/router": "10.2.2",
+    "@angular/animations": "11.2.6",
+    "@angular/cdk": "11.2.6",
+    "@angular/common": "11.2.6",
+    "@angular/compiler": "11.2.6",
+    "@angular/core": "11.2.6",
+    "@angular/forms": "11.2.6",
+    "@angular/localize": "11.2.6",
+    "@angular/material": "11.2.6",
+    "@angular/platform-browser": "11.2.6",
+    "@angular/platform-browser-dynamic": "11.2.6",
+    "@angular/router": "11.2.6",
+    "@auth0/auth0-angular": "1.4.0",
+    "@ng-bootstrap/ng-bootstrap": "9.0.2",
+    "@ngxs/store": "3.7.1",
     "angular-in-memory-web-api": "0.11.0",
-    "@auth0/auth0-angular": "1.2.0",
     "body-parser": "1.19.0",
-    "bootstrap": "4.4.1",
-    "core-js": "3.6.4",
-    "debug": "4.1.1",
+    "bootstrap": "4.6.0",
+    "core-js": "3.9.1",
+    "debug": "4.3.1",
     "express": "4.17.1",
     "express-jwt": "6.0.0",
-    "fast-sort": "2.0.1",
-    "jquery": "3.5.0",
-    "jwks-rsa": "1.11.0",
-    "moment": "2.24.0",
-    "morgan": "1.9.1",
-    "@ng-bootstrap/ng-bootstrap": "8.0.0",
+    "fast-sort": "2.2.0",
+    "immer": "9.0.1",
+    "jquery": "3.6.0",
+    "jwks-rsa": "2.0.1",
+    "moment": "2.29.1",
+    "morgan": "1.10.0",
     "path": "0.12.7",
-    "pg": "8.4.2",
-    "popper.js": "1.16.0",
-    "rxjs": "6.6.3",
-    "sequelize": "6.3.5",
+    "pg": "8.5.1",
+    "popper.js": "1.16.1",
+    "rxjs": "6.6.6",
+    "sequelize": "6.5.1",
     "socket.io": "2.4.0",
     "socket.io-client": "2.3.0",
-    "typescript": "4.0.5",
-    "underscore": "1.9.2",
+    "typescript": "4.1.5",
+    "underscore": "1.12.1",
     "yargs": "16.2.0",
-    "zone.js": "0.10.3"
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1002.0",
-    "@angular/cli": "10.2.0",
-    "@angular/compiler-cli": "10.2.2",
-    "@angular/language-service": "10.2.2",
-    "@types/jasmine": "3.5.5",
+    "@angular-devkit/build-angular": "0.1102.5",
+    "@angular/cli": "11.2.5",
+    "@angular/compiler-cli": "11.2.6",
+    "@angular/language-service": "11.2.6",
+    "@ngxs/devtools-plugin": "3.7.1",
+    "@ngxs/logger-plugin": "3.7.1",
+    "@types/jasmine": "3.6.7",
     "@types/jasminewd2": "2.0.8",
-    "@types/node": "13.1.7",
+    "@types/node": "14.14.35",
+    "@types/underscore": "1.10.24",
     "codelyzer": "6.0.1",
-    "enhanced-resolve": "4.1.1",
-    "jasmine-core": "3.5.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "5.1.1",
+    "enhanced-resolve": "5.7.0",
+    "jasmine-core": "3.7.1",
+    "jasmine-spec-reporter": "6.0.0",
+    "karma": "6.2.0",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-jasmine": "3.1.0",
-    "karma-jasmine-html-reporter": "1.5.1",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.5.4",
     "protractor": "7.0.0",
-    "ts-node": "8.6.2",
-    "tslib": "1.10.0",
-    "tslint": "5.20.1"
+    "ts-node": "9.1.1",
+    "tslib": "2.1.0",
+    "tslint": "6.1.3"
   },
   "engines": {
     "node": "14.15.0",
-    "npm": "6.14.8"
+    "npm": "7.7.5"
   }
 }
