@@ -1,66 +1,32 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {FinalResult} from '../interfaces/FinalResult';
-import {Observable, of, Subscription} from 'rxjs';
-import {catchError} from 'rxjs/operators';
-import {_} from 'underscore';
+import {Observable} from 'rxjs';
+import {filter, map} from 'rxjs/operators';
+import * as _ from 'underscore';
+import {Store} from '@ngxs/store';
+import {GetFinalResults} from '../actions/final-result.action';
 
 @Injectable({
   providedIn: 'root'
 })
 export class FinalResultsService {
-  finalResultsUrl = 'api/finalResults';
-  private finalResults: FinalResult[];
 
-  constructor(private http: HttpClient) {
-    this.getFinalResults().subscribe();
-  }
+  finalResults$: Observable<FinalResult[]> = this.store.select(state => state.finalResults).pipe(
+    filter(model => !!model),
+    map(model => model.finalResults),
+    filter(finalResults => !!finalResults)
+  );
 
-  public getFinalResults(): Observable<FinalResult[]> {
-    if (!!this.finalResults) {
-      return of(this.finalResults);
-    } else {
-      return new Observable<FinalResult[]>(observer => {
-        this.http.get<FinalResult[]>(this.finalResultsUrl)
-          .pipe(
-            catchError(this.handleError<FinalResult[]>('getFinalResults', []))
-          )
-          .subscribe((finalResults: FinalResult[]) => {
-            this.finalResults = finalResults;
-            observer.next(finalResults);
-          });
-      });
-    }
+  constructor(private http: HttpClient,
+              private store: Store) {
+    this.store.dispatch(new GetFinalResults());
   }
 
   public getFinalResultsForGroup(group_id: number): Observable<FinalResult[]> {
-    return new Observable<FinalResult[]>(observer => {
-      this.getFinalResults().subscribe(results => {
-        observer.next(_.filter(results, result => result.group_id === group_id));
-      });
-    });
-  }
-
-  stillLoading(): boolean {
-    return this.finalResults.length === 0;
-  }
-
-
-  /**
-   * Handle Http operation that failed.
-   * Let the app continue.
-   * @param operation - name of the operation that failed
-   * @param result - optional value to return as the observable result
-   */
-  private handleError<T> (operation = 'operation', result?: T) {
-    return (error: any): Observable<T> => {
-
-      // TODO: send the error to remote logging infrastructure
-      console.error(error); // log to console instead
-
-      // Let the app keep running by returning an empty result.
-      return of(result as T);
-    };
+    return this.finalResults$.pipe(
+      map(finalResults => _.where(finalResults, {group_id}))
+    );
   }
 
 }
