@@ -1,24 +1,22 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {SystemVarsService} from '../../services/system.vars.service';
 import {CategoryService} from '../../services/category.service';
 import {VotesService} from '../../services/votes.service';
-import {_} from 'underscore';
 import {WinnersService} from '../../services/winners.service';
-import {MyAuthService} from '../../services/auth/my-auth.service';
-import {OddsService} from '../../services/odds.service';
+import {Observable} from 'rxjs';
+import {first, map} from 'rxjs/operators';
 import {PersonService} from '../../services/person.service';
+import fast_sort from 'fast-sort';
+import {SocketService} from '../../services/socket.service';
 
 @Component({
   selector: 'osc-admin-dashboard',
   templateUrl: './admin-dashboard.component.html',
-  styleUrls: ['./admin-dashboard.component.scss']
+  styleUrls: ['./admin-dashboard.component.scss'],
 })
 export class AdminDashboardComponent implements OnInit {
 
   reloadingData = false;
-
-  currentYear: number;
-  possibleYears: number[] = [];
 
   winnersDeleting = false;
   winnersDeleted = false;
@@ -27,77 +25,75 @@ export class AdminDashboardComponent implements OnInit {
               private categoryService: CategoryService,
               private votesService: VotesService,
               private winnersService: WinnersService,
-              private oddsService: OddsService,
-              public auth: MyAuthService) { }
+              private socket: SocketService,
+              private personService: PersonService) { }
 
-  ngOnInit() {
-    this.systemVarsService.getSystemVars().subscribe(systemVars => {
-      this.categoryService.getMaxYear().subscribe(maxYear => {
-        this.currentYear = systemVars.curr_year;
-        this.possibleYears.push(maxYear - 1);
-        this.possibleYears.push(maxYear);
-        if (!_.contains(this.possibleYears, this.currentYear)) {
-          this.possibleYears.push(this.currentYear);
-        }
-        this.categoryService.subscribeToWinnerEvents().subscribe(() => {
-          this.winnersDeleting = false;
-          this.winnersDeleted = true;
-        });
-      });
+  ngOnInit(): void {
+    this.socket.on('reset_winners', () => {
+      this.winnersDeleting = false;
+      this.winnersDeleted = true;
     });
   }
 
-  refreshAllData(): void {
-    this.categoryService.refreshCache().subscribe();
-    this.votesService.refreshCacheForThisYear().subscribe();
-    this.oddsService.refreshCache().subscribe();
+  get isAdmin(): boolean {
+    return this.personService.isAdmin;
+  }
+
+  get currentYear(): Observable<number> {
+    return this.systemVarsService.getCurrentYear();
+  }
+
+  getPossibleYears(currentYear: number): Observable<number[]> {
+    return this.categoryService.getMaxYear().pipe(
+      map(maxYear => {
+        const yearsSet = new Set<number>();
+        yearsSet.add(maxYear);
+        yearsSet.add(maxYear - 1);
+        yearsSet.add(currentYear);
+        const yearsList = Array.from(yearsSet);
+        fast_sort(yearsList);
+        return yearsList;
+      })
+    );
   }
 
   stillLoading(): boolean {
-    return this.categoryService.stillLoading() ||
-      this.systemVarsService.stillLoading();
+    return this.systemVarsService.stillLoading();
   }
 
-  isVotingOpen(): boolean {
+  isVotingOpen(): Observable<boolean> {
     return this.systemVarsService.canVote();
   }
 
-  getVotingButtonClass(votingOpen: boolean): string {
-    return this.isVotingOpen() === votingOpen ? 'btn-success' : 'btn-primary';
-  }
-
-  yearButtonClass(year): string {
-    if (this.currentYear === year) {
-      return 'btn-success';
-    } else {
-      return 'btn-primary';
-    }
+  getVotingButtonClass(votingOpen: boolean): Observable<string> {
+    return this.isVotingOpen().pipe(
+      map(isVotingOpen => isVotingOpen === votingOpen ? 'btn-success' : 'btn-primary')
+    );
   }
 
   changeCurrentYear(year): void {
     this.reloadingData = true;
-    this.systemVarsService.changeCurrentYear(year).subscribe(() => {
-      this.categoryService.refreshCache().subscribe(() => {
-        this.votesService.refreshCache(year).subscribe(() => {
-          this.reloadingData = false;
-          this.currentYear = year;
-        });
-      });
-    });
+    this.systemVarsService.changeCurrentYear(year);
   }
 
   toggleVotingLock(votingOpen: boolean): void {
-    if (votingOpen !== this.isVotingOpen()) {
-      this.systemVarsService.toggleVotingLock();
-    }
+    this.isVotingOpen().pipe(first())
+      .subscribe(isVotingOpen => {
+        if (votingOpen !== isVotingOpen) {
+          this.systemVarsService.toggleVotingLock();
+        }
+      }
+    );
   }
 
   resetWinners(): void {
-    this.systemVarsService.getSystemVars().subscribe(systemVars => {
-      const year = systemVars.curr_year;
-      this.winnersDeleting = true;
-      this.winnersService.resetWinners(year).subscribe();
-    });
+    this.systemVarsService.systemVars
+      .pipe(first())
+      .subscribe(systemVars => {
+        const year = systemVars.curr_year;
+        this.winnersDeleting = true;
+        this.winnersService.resetWinners(year);
+      });
   }
 
   getWinnersButtonClass(): string {
