@@ -10,7 +10,7 @@ import {CategoryService} from '../../services/category.service';
 import {CategoryServiceStub} from '../../services/category.service.stub';
 import {TestCategoryList} from '../../services/data/categories.test.mock';
 import {Category} from '../../interfaces/Category';
-import {_} from 'underscore';
+import * as _ from 'underscore';
 import {DebugElement} from '@angular/core';
 import {By} from '@angular/platform-browser';
 import {CallbackComponent} from '../callback/callback.component';
@@ -23,6 +23,7 @@ import {ActiveContext} from '../categories.context';
 import {WinnerMainComponent} from '../winner-main/winner-main.component';
 import {WinnerDetailComponent} from '../winner-detail/winner-detail.component';
 import {ScoreboardComponent} from '../scoreboard/scoreboard.component';
+import {combineLatest, Observable, of} from 'rxjs';
 
 describe('CategoryHopperComponent', () => {
   let component: CategoryHopperComponent;
@@ -62,10 +63,8 @@ describe('CategoryHopperComponent', () => {
     service = TestBed.get(CategoryService);
   });
 
-  function populateInputs(categoryIndex: number) {
-    const currentCategory = getCategory(categoryIndex);
-    component.nominees = currentCategory ? currentCategory.nominees : [];
-
+  function populateInputs(categoryIndex: number): void {
+    component.category = getCategory(categoryIndex);
     component.next = getCategory(categoryIndex + 1);
     component.prev = getCategory(categoryIndex - 1);
     component.activeContext = ActiveContext.OddsAssignment;
@@ -73,11 +72,11 @@ describe('CategoryHopperComponent', () => {
     fixture.detectChanges();
   }
 
-  function getCategory(categoryIndex): Category {
-    return categoryIndex >= TestCategoryList.length ? null : TestCategoryList[categoryIndex];
+  function getCategory(categoryIndex): Observable<Category> {
+    return categoryIndex >= TestCategoryList.length ? of(null) : of(TestCategoryList[categoryIndex]);
   }
 
-  function findButtonWithText(subElement: DebugElement, textToFind: String): DebugElement {
+  function findButtonWithText(subElement: DebugElement, textToFind: string): DebugElement {
     const debugElements = subElement.queryAll(By.css('button'));
     return _.find(debugElements, (button) => {
       return button.nativeElement.innerHTML.includes(textToFind);
@@ -96,25 +95,40 @@ describe('CategoryHopperComponent', () => {
     expect(component.prev).toBeFalsy();
   });
 
-  it('inputs populated if all exist', () => {
+  it('inputs populated if all exist', done => {
     populateInputs(1);
-    expect(component.nominees.length).toBe(3);
-    expect(component.next).toBeTruthy();
-    expect(component.prev).toBeTruthy();
+    combineLatest([component.category, component.next, component.prev]).subscribe(
+      ([category, next, prev]) => {
+        expect(category.nominees.length).toBe(3);
+        expect(next).toBeTruthy();
+        expect(prev).toBeTruthy();
+        done();
+      }
+    );
   });
 
-  it('next empty if last', () => {
+  it('next empty if last', done => {
     populateInputs(2);
-    expect(component.nominees.length).toBe(3);
-    expect(component.next).toBeFalsy();
-    expect(component.prev).toBeTruthy();
+    combineLatest([component.category, component.next, component.prev]).subscribe(
+      ([category, next, prev]) => {
+        expect(category.nominees.length).toBe(3);
+        expect(next).toBeFalsy();
+        expect(prev).toBeTruthy();
+        done();
+      }
+    );
   });
 
-  it('prev empty if first', () => {
+  it('prev empty if first', done => {
     populateInputs(0);
-    expect(component.nominees.length).toBe(3);
-    expect(component.next).toBeTruthy();
-    expect(component.prev).toBeFalsy();
+    combineLatest([component.category, component.next, component.prev]).subscribe(
+      ([category, next, prev]) => {
+        expect(category.nominees.length).toBe(3);
+        expect(next).toBeTruthy();
+        expect(prev).toBeFalsy();
+        done();
+      }
+    );
   });
 
   it('all buttons exist for all inputs', () => {
@@ -148,9 +162,12 @@ describe('CategoryHopperComponent', () => {
     expect(component.totalOdds('user')).toBe(81);
   });
 
-  it('totalOddsVegas', () => {
+  it('totalOddsVegas', done => {
     populateInputs(1);
-    expect(component.totalOddsVegas()).toBe(97.06959706959707);
+    component.totalOddsVegas().subscribe(totalOddsVegas => {
+      expect(totalOddsVegas).toBe(97.06959706959707);
+      done();
+    });
   });
 
   it('totalOdds unrecognized', () => {
@@ -158,9 +175,12 @@ describe('CategoryHopperComponent', () => {
     expect(component.totalOdds('unrecognized')).toBe(0);
   });
 
-  it('hasChanges is false after inputs', () => {
+  it('hasChanges is false after inputs', done => {
     populateInputs(1);
-    expect(component.hasChanges()).toBe(false);
+    component.hasChanges().subscribe(hasChanges => {
+      expect(hasChanges).toBe(false);
+      done();
+    });
   });
 
   it('submit button disabled on init', () => {
@@ -170,96 +190,56 @@ describe('CategoryHopperComponent', () => {
       .toBeTruthy('submit button is enabled');
   });
 
-  it('hasChanges is true after one change', () => {
+  it('hasChanges is true after one change', done => {
+    populateInputs(1);
+    component.category.subscribe(category => {
+      const originalValue = category.nominees[0].odds_expert;
+
+      category.nominees[0].odds_expert = 23;
+      component.hasChanges().subscribe(hasChanges => {
+        expect(hasChanges).toBe(true);
+
+        // reset to initial values for later test runs
+        category.nominees[0].odds_expert = originalValue;
+        done();
+      });
+
+    });
+  });
+
+  it('hasChanges is true after two different changes', done => {
+    populateInputs(1);
+    const originalFirst = component.nominees[0].odds_expert;
+    const originalSecond = component.nominees[1].odds_expert;
+
+    component.nominees[0].odds_expert = 23;
+    component.nominees[1].odds_expert = 3;
+
+    component.hasChanges().subscribe(hasChanges => {
+      expect(hasChanges).toBe(true);
+
+      // reset to initial values for later test runs
+      component.nominees[0].odds_expert = originalFirst;
+      component.nominees[1].odds_expert = originalSecond;
+
+      done();
+    });
+  });
+
+  it('hasChanges is false after reverting change', done => {
     populateInputs(1);
     const originalValue = component.nominees[0].odds_expert;
 
     component.nominees[0].odds_expert = 23;
-    expect(component.hasChanges()).toBe(true);
-
-    // reset to initial values for later test runs
-    component.nominees[0].odds_expert = originalValue;
-  });
-
-  it('hasChanges is true after two different changes', () => {
-    populateInputs(1);
-    const originalFirst = component.nominees[0].odds_expert;
-    const originalSecond = component.nominees[1].odds_expert;
-
-    component.nominees[0].odds_expert = 23;
-    component.nominees[1].odds_expert = 3;
-    expect(component.hasChanges()).toBe(true);
-
-    // reset to initial values for later test runs
-    component.nominees[0].odds_expert = originalFirst;
-    component.nominees[1].odds_expert = originalSecond;
-  });
-
-  it('hasChanges is false after reverting change', () => {
-    populateInputs(1);
-    const originalValue = component.nominees[0].odds_expert;
-
-    component.nominees[0].odds_expert = 23;
     component.nominees[0].odds_expert = originalValue;
 
-    expect(component.hasChanges()).toBe(false);
+    component.hasChanges().subscribe(hasChanges => {
+      expect(hasChanges).toBe(false);
+      done();
+    });
   });
 
-  it('submitOdds calls updateNominee for one change', () => {
-    populateInputs(1);
-    const originalFirst = component.nominees[0].odds_expert;
-
-    component.nominees[0].odds_expert = 23;
-
-    const updateSpy = spyOn(service, 'updateNominee').and.callThrough();
-
-    component.submitOdds();
-
-    expect(updateSpy).toHaveBeenCalledTimes(1);
-
-    // reset to initial values for later test runs
-    component.nominees[0].odds_expert = originalFirst;
-  });
-
-  it('submitOdds calls updateNominee twice for two changes', () => {
-    populateInputs(1);
-    const originalFirst = component.nominees[0].odds_expert;
-    const originalSecond = component.nominees[1].odds_expert;
-
-    component.nominees[0].odds_expert = 23;
-    component.nominees[1].odds_expert = 3;
-
-    const updateSpy = spyOn(service, 'updateNominee').and.callThrough();
-
-    component.submitOdds();
-
-    expect(updateSpy).toHaveBeenCalledTimes(2);
-
-    // reset to initial values for later test runs
-    component.nominees[0].odds_expert = originalFirst;
-    component.nominees[1].odds_expert = originalSecond;
-  });
-
-  it('submitOdds calls updateNominee once for two changes on one nominee', () => {
-    populateInputs(1);
-    const originalFirst = component.nominees[0].odds_expert;
-    const originalSecond = component.nominees[0].odds_user;
-
-    component.nominees[0].odds_expert = 23;
-    component.nominees[0].odds_user = 3;
-
-    const updateSpy = spyOn(service, 'updateNominee').and.callThrough();
-
-    component.submitOdds();
-
-    expect(updateSpy).toHaveBeenCalledTimes(1);
-
-    // reset to initial values for later test runs
-    component.nominees[0].odds_expert = originalFirst;
-    component.nominees[0].odds_user = originalSecond;
-  });
-
-  it('hasChanges is false after submitOdds is called', () => {
+  it('hasChanges is false after submitOdds is called', done => {
     populateInputs(1);
     const originalFirst = component.nominees[0].odds_expert;
     const originalSecond = component.nominees[1].odds_expert;
@@ -269,11 +249,15 @@ describe('CategoryHopperComponent', () => {
 
     component.submitOdds();
 
-    expect(component.hasChanges()).toBe(false);
+    component.hasChanges().subscribe(hasChanges => {
+      expect(hasChanges).toBe(false);
 
-    // reset to initial values for later test runs
-    component.nominees[0].odds_expert = originalFirst;
-    component.nominees[1].odds_expert = originalSecond;
+      // reset to initial values for later test runs
+      component.nominees[0].odds_expert = originalFirst;
+      component.nominees[1].odds_expert = originalSecond;
+
+      done();
+    });
   });
 
 });
