@@ -1,13 +1,15 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {Category} from '../../interfaces/Category';
 import {Nominee} from '../../interfaces/Nominee';
-import {_} from 'underscore';
-import {forkJoin, Observable} from 'rxjs';
+import * as _ from 'underscore';
+import {Observable} from 'rxjs';
 import {CategoryService} from '../../services/category.service';
 import {ActiveContext} from '../categories.context';
 import {VotesService} from '../../services/votes.service';
+import {map, mergeMap} from 'rxjs/operators';
 import {PersonService} from '../../services/person.service';
-import {MyAuthService} from '../../services/auth/my-auth.service';
+import {NomineeControls} from '../nominees/nominees.component';
+import {OddsChange} from '../../actions/category.action';
 
 @Component({
   selector: 'osc-category-hopper',
@@ -16,28 +18,33 @@ import {MyAuthService} from '../../services/auth/my-auth.service';
 })
 export class CategoryHopperComponent implements OnInit {
 
-  @Input() next: Category;
-  @Input() prev: Category;
-  @Input() nominees: Nominee[];
+  @Input() next: Observable<Category>;
+  @Input() prev: Observable<Category>;
+  @Input() category: Observable<Category>;
   @Input() activeContext: ActiveContext;
+  @Input() nomineeGroups: NomineeControls[];
+
   private readonly contextUrls: string[];
-  categoryCount: number;
+
+  categoryCount = this.categoryService.categories.pipe(
+    map(categories => categories.length)
+  );
+
+  nominees$: Observable<Nominee[]>;
 
   constructor(private categoryService: CategoryService,
               private votesService: VotesService,
-              private auth: MyAuthService) {
+              private personService: PersonService) {
     this.contextUrls = [];
     this.contextUrls[ActiveContext.Vote] = 'vote';
     this.contextUrls[ActiveContext.OddsAssignment] = 'odds';
     this.contextUrls[ActiveContext.Winner] = 'winners';
-
-    this.categoryService.getCategories().subscribe(categories => {
-      this.categoryCount = categories.length;
-    });
   }
 
-  ngOnInit() {
-    this.clearOriginals();
+  ngOnInit(): void {
+    this.nominees$ = this.category.pipe(
+      map(category => category.nominees)
+    );
   }
 
   baseLink(): string {
@@ -48,34 +55,36 @@ export class CategoryHopperComponent implements OnInit {
     return ActiveContext.Vote === this.activeContext;
   }
 
-  numVotesComplete(): number {
-    const me = this.auth.getPersonNow();
-    if (!!me) {
-      return this.votesService.getVotesForCurrentYearAndPerson(me).length;
-    } else {
-      return 0;
-    }
-  }
-
-  percentVotesComplete(): number {
-    const voteCount = this.numVotesComplete();
-    return !!this.categoryCount ? (voteCount * 100) / this.categoryCount : 0;
+  numVotesComplete(): Observable<number> {
+    return this.personService.me$.pipe(
+      mergeMap(me => this.votesService.getVotesForCurrentYearAndPerson(me)),
+      map(votes => votes.length)
+    );
   }
 
   showOdds(): boolean {
     return ActiveContext.OddsAssignment === this.activeContext;
   }
 
-  totalOdds(subtitle: string): number {
-    const odds_nums = _.map(this.nominees, (nominee) => nominee[`odds_${subtitle}`]);
-    return this.oddsCalc(odds_nums);
+  totalOdds(subtitle: string): Observable<number> {
+    return this.nominees$.pipe(
+      map(nominees => {
+        const odds_nums = _.map(nominees, (nominee) => +nominee[`odds_${subtitle}`]);
+        return this.oddsCalc(odds_nums);
+      })
+    );
   }
 
-  totalOddsVegas(): number {
-    const odds_nums = _.map(this.nominees, (nominee) => this.vegasCalc(nominee.odds_numerator, nominee.odds_denominator));
-    return this.oddsCalc(odds_nums);
+  totalOddsVegas(): Observable<number> {
+    return this.nominees$.pipe(
+      map(nominees => {
+        const odds_nums = _.map(nominees, (nominee) => this.vegasCalc(nominee.odds_numerator, nominee.odds_denominator));
+        return this.oddsCalc(odds_nums);
+      })
+    );
   }
 
+  // noinspection JSMethodCanBeStatic
   private vegasCalc(numerator?: number, denominator?: number): number {
     if (!numerator || !denominator) {
       return 0;
@@ -84,44 +93,57 @@ export class CategoryHopperComponent implements OnInit {
     }
   }
 
+  // noinspection JSMethodCanBeStatic
   private oddsCalc(odds_nums: number[]): number {
     const compacted = _.compact(odds_nums);
     const total_num = _.reduce(compacted, (memo, later) => memo + later);
     return total_num ? total_num : 0;
   }
 
-  private getChanges() {
-    return _.filter(this.nominees, (nominee) =>
-      nominee.original_odds_expert !== nominee.odds_expert ||
-      nominee.original_odds_user !== nominee.odds_user ||
-      nominee.original_odds_numerator !== nominee.odds_numerator ||
-      nominee.original_odds_denominator !== nominee.odds_denominator
-    );
+  private getChanges(): OddsChange[] {
+    const changes: OddsChange[] = [];
+    _.each(this.nomineeGroups, ng => {
+      if (this.isChanged(ng)) {
+        changes.push(this.createOddsChange(ng));
+      }
+    });
+    return changes;
   }
 
-  hasChanges() {
-    const filtered = this.getChanges();
-    return filtered.length > 0;
+  hasChanges(): boolean {
+    return _.filter(this.nomineeGroups, ng => this.isChanged(ng)).length > 0;
   }
 
-  submitOdds() {
-    const requests: Observable<any>[] = [];
+  submitOdds(): void {
     const changes = this.getChanges();
-    const categoryService = this.categoryService;
-    _.forEach(changes, function(nominee) {
-      requests.push(categoryService.updateNominee(nominee));
-    });
-    forkJoin(requests).subscribe(() => {
-      this.clearOriginals();
-    });
+    this.categoryService.updateOddsForNominees(changes).subscribe(() => this.clearOriginals());
   }
 
-  private clearOriginals() {
-    _.forEach(this.nominees, (nominee) => {
-      nominee.original_odds_expert = nominee.odds_expert;
-      nominee.original_odds_user = nominee.odds_user;
-      nominee.original_odds_numerator = nominee.odds_numerator;
-      nominee.original_odds_denominator = nominee.odds_denominator;
+  // noinspection JSMethodCanBeStatic
+  private createOddsChange(nomineeControls: NomineeControls): OddsChange {
+    return {
+      nomination_id: nomineeControls.nominee.id,
+      odds_expert: nomineeControls.expert.value,
+      odds_user: nomineeControls.user.value,
+      odds_numerator: nomineeControls.numerator.value,
+      odds_denominator: nomineeControls.denominator.value,
+    };
+  }
+
+  // noinspection JSMethodCanBeStatic
+  private isChanged(nomineeControl: NomineeControls): boolean {
+    return nomineeControl.expert.dirty ||
+      nomineeControl.user.dirty ||
+      nomineeControl.numerator.dirty ||
+      nomineeControl.denominator.dirty;
+  }
+
+  private clearOriginals(): void {
+    _.each(this.nomineeGroups, ng => {
+      ng.expert.reset(ng.expert.value);
+      ng.user.reset(ng.user.value);
+      ng.numerator.reset(ng.numerator.value);
+      ng.denominator.reset(ng.denominator.value);
     });
   }
 
