@@ -1,4 +1,4 @@
-export class SystemVars {
+export interface SystemVars {
   id?: number;
   curr_year: number;
   voting_open: boolean;
