@@ -3,7 +3,7 @@ import {SystemVars} from '../../interfaces/SystemVars';
 export const MockSystemVars: SystemVars[] = [
   {
     id: 1,
-    curr_year: 2019,
-    voting_open: true
+    curr_year: 2018,
+    voting_open: false
   }
 ];
