@@ -1,12 +1,11 @@
 import {getStatusText, InMemoryDbService, RequestInfo, ResponseOptions, STATUS} from 'angular-in-memory-web-api';
 import {Injectable} from '@angular/core';
-import {_} from 'underscore';
+import * as _ from 'underscore';
 import {MockCategoryList} from './data/categories.mock';
 import {MockPersonList} from './data/persons.mock';
 import {MockVoteList} from './data/votes.mock';
 import {Vote} from '../interfaces/Vote';
-import {Event} from '../interfaces/Event';
-import {Observable, of} from 'rxjs';
+import {Observable} from 'rxjs';
 import {MockSystemVars} from './data/system.vars.mock';
 import {MockWinnerList} from './data/winners.mock';
 import {MockEvents} from './data/event.mock';
@@ -14,6 +13,8 @@ import {MockOdds} from './data/odds.mock';
 import {MockFinalResultsList} from './data/finalresults.mock';
 import {Category} from '../interfaces/Category';
 import {InMemoryCallbacksService} from './in-memory-callbacks.service';
+import {ArrayUtil} from '../utility/ArrayUtil';
+import {Nominee} from '../interfaces/Nominee';
 
 @Injectable({
   providedIn: 'root',
@@ -32,7 +33,7 @@ export class InMemoryDataService implements InMemoryDbService {
 
   /////////// helpers ///////////////
 
-  private static finishOptions(options: ResponseOptions, {headers, url}: RequestInfo) {
+  private static finishOptions(options: ResponseOptions, {headers, url}: RequestInfo): ResponseOptions {
     options.statusText = getStatusText(options.status);
     options.headers = headers;
     options.url = url;
@@ -54,23 +55,32 @@ export class InMemoryDataService implements InMemoryDbService {
       winners: this.winners,
       events: this.events,
       odds: this.odds,
-      finalResults: this.finalResults
+      finalResults: this.finalResults,
+      oddsChange: [],
+      resetWinners: [],
     };
   }
 
-  on(channel, callback) {
+  /* SOCKET METHODS */
+
+  on(channel, callback): void {
     this.callbackService.on(channel, callback);
   }
 
-  removeCallback(channel, callback) {
+  removeCallback(channel, callback): void {
     this.callbackService.removeCallback(channel, callback);
   }
 
-  getCallbacks(channel) {
+  getCallbacks(channel): any[] {
     return this.callbackService.getCallbacks(channel);
   }
 
-  get(requestInfo: RequestInfo) {
+  broadcastToChannel(channel, msg): void {
+    const callbacks = this.getCallbacks(channel);
+    _.forEach(callbacks, callback => callback(msg));
+  }
+
+  get(requestInfo: RequestInfo): Observable<Response> {
     const collectionName = requestInfo.collectionName;
     if (collectionName === 'categories') {
       return this.getCategoriesWithVotes(requestInfo);
@@ -82,100 +92,186 @@ export class InMemoryDataService implements InMemoryDbService {
   }
 
   // noinspection JSUnusedGlobalSymbols
-  put(requestInfo: RequestInfo) {
+  put(requestInfo: RequestInfo): Observable<Response> {
     console.log('HTTP override: PUT');
     const collectionName = requestInfo.collectionName;
     if (collectionName === 'nominees') {
       this.updateNomination(requestInfo);
-    } else if (requestInfo.collectionName === 'systemVars') {
+    } else if (collectionName === 'systemVars') {
       this.updateSystemVars(requestInfo);
+    } else if (collectionName === 'oddsChange') {
+      return this.updateOdds(requestInfo);
+    } else if (collectionName === 'resetWinners') {
+      return this.deleteWinners(requestInfo);
     }
     return undefined;
   }
 
-  // noinspection JSUnusedGlobalSymbols
-  patch(requestInfo: RequestInfo) {
-    console.log('HTTP override: PATCH');
-    const collectionName = requestInfo.collectionName;
-    if (collectionName === 'winners') {
-      this.deleteWinners();
-    }
-    return undefined;
-  }
-
-  post(requestInfo: RequestInfo) {
+  post(requestInfo: RequestInfo): Observable<Response> {
     if (requestInfo.collectionName === 'votes') {
-      const existingVote = this.existingVote(requestInfo);
-      if (existingVote) {
-        return this.updateVote(requestInfo, existingVote);
-      }
+      return this.addOrChangeVote(requestInfo);
     } else if (requestInfo.collectionName === 'winners') {
-      this.addOrDeleteWinner(requestInfo);
+      this.addWinner(requestInfo);
     }
     return undefined;
   }
 
-  private deleteWinners() {
-    _.forEach(this.categories, category => category.winners = []);
+  delete(requestInfo: RequestInfo): Observable<Response> {
+    if (requestInfo.collectionName === 'winners') {
+      this.removeWinner(requestInfo);
+    }
+    return undefined;
+  }
+
+  private deleteWinners(requestInfo: RequestInfo): Observable<Response> {
+    const year = this.getBody(requestInfo).year;
+
+    _.forEach(this.categories, category => {
+      const winners = ArrayUtil.cloneArray(category.winners);
+      _.forEach(winners, (winner: any) => {
+        if (winner.year === year) {
+          ArrayUtil.removeFromArray(winners, winner);
+        }
+      });
+    });
 
     const socketMsg = {
-      detail: 'reset',
       event_id: 1,
-      event_time: new Date
+      event_time: new Date(),
+      year
     };
-    const callbacks = this.getCallbacks('winner');
-    _.forEach(callbacks, callback => callback(socketMsg));
+
+    this.broadcastToChannel('reset_winners', socketMsg);
+
+    this.sendUpdatedOdds();
+
+    return this.packageUpResponse({msg: 'Success!'}, requestInfo);
+  }
+
+  private updateSystemVars(requestInfo: RequestInfo): void {
+    const jsonBody = requestInfo.utils.getJsonBody(requestInfo.req);
+    const systemVars = this.systemVars[0];
+
+    if (jsonBody.voting_open !== undefined) {
+      systemVars.voting_open = jsonBody.voting_open;
+
+      const msg = {
+        event_id: 1,
+        event_time: new Date()
+      };
+
+      if (!systemVars.voting_open) {
+        this.broadcastToChannel('voting_locked', msg);
+      } else {
+        this.broadcastToChannel('voting_unlocked', msg);
+      }
+    }
+  }
+
+  private addOrChangeVote(requestInfo: RequestInfo): Observable<Response> {
+    const body = this.getBody(requestInfo);
+
+    const existing = _.findWhere(this.votes, {
+      category_id: body.category_id,
+      year: body.year,
+      person_id: body.person_id
+    });
+
+    if (!!existing) {
+      existing.nomination_id = body.nomination_id;
+      const msg = {
+        vote_id: existing.id,
+        nomination_id: body.nomination_id
+      };
+      this.broadcastToChannel('change_vote', msg);
+
+      return this.packageUpResponse({msg: 'Success!'}, requestInfo);
+
+    } else {
+      const vote_id = this.genVoteId();
+      const vote = {
+        id: vote_id,
+        category_id: body.category_id,
+        year: body.year,
+        person_id: body.person_id,
+        nomination_id: body.nomination_id
+      };
+      this.votes.push(vote);
+      this.broadcastToChannel('add_vote', vote);
+
+      return this.packageUpResponse(vote, requestInfo);
+    }
+  }
+
+  private addWinner(requestInfo: RequestInfo): void {
+    const jsonBody = requestInfo.utils.getJsonBody(requestInfo.req);
+
+    let socketMsg;
+
+    const categoryForNominee = this.categoryForNominee(jsonBody.nomination_id);
+    jsonBody.id = this.genWinnerId();
+    categoryForNominee.winners.push(jsonBody);
+    socketMsg = {
+      nomination_id: jsonBody.nomination_id,
+      event_id: 1,
+      event_time: new Date(),
+      winner_id: jsonBody.id,
+      declared: new Date(),
+    };
+
+    this.broadcastToChannel('add_winner', socketMsg);
 
     this.sendUpdatedOdds();
   }
 
-  private updateSystemVars(requestInfo: RequestInfo) {
-    const jsonBody = requestInfo.utils.getJsonBody(requestInfo.req);
-    const systemVars = this.systemVars[0];
-
-    if (systemVars.voting_open !== jsonBody.voting_open) {
-      systemVars.voting_open = jsonBody.voting_open;
-
-      const msg = {
-        voting_open: systemVars.voting_open,
-        event_id: 1,
-        event_time: new Date
-      };
-
-      const callbacks = this.getCallbacks('voting');
-      _.forEach(callbacks, callback => callback(msg));
-    }
+  private logReadOnly(): void {
+    let readOnly = 0;
+    let writeable = 0;
+    _.chain(this.categories)
+      .map(c => c.nominees)
+      .flatten()
+      .each(n => {
+        if (Object.getOwnPropertyDescriptor(n, 'nominee').writable) {
+          writeable++;
+        } else {
+          readOnly++;
+        }
+      });
+    console.log(`Read-only: ${readOnly}, Writeable: ${writeable}`);
   }
 
-  private addOrDeleteWinner(requestInfo: RequestInfo) {
+  private updateOdds(requestInfo: RequestInfo): Observable<Response> {
     const jsonBody = requestInfo.utils.getJsonBody(requestInfo.req);
-    const categoryForExistingWinner = this.existingCategoryForWinner(jsonBody.nomination_id);
+    const changes = jsonBody.changes;
 
-    let socketMsg;
-    if (!!categoryForExistingWinner) {
-      const existingWinner = _.findWhere(categoryForExistingWinner.winners, {nomination_id: jsonBody.nomination_id});
-      categoryForExistingWinner.winners = _.without(categoryForExistingWinner.winners, existingWinner);
-      socketMsg = {
-        detail: 'delete',
-        nomination_id: jsonBody.nomination_id,
-        event_id: 1,
-        event_time: new Date
-      };
-    } else {
-      const categoryForNominee = this.categoryForNominee(jsonBody.nomination_id);
-      categoryForNominee.winners.push(jsonBody);
-      socketMsg = {
-        detail: 'add',
-        nomination_id: jsonBody.nomination_id,
-        event_id: 1,
-        event_time: new Date,
-        winner_id: 1234,
-        declared: new Date,
-      };
-    }
+    this.logReadOnly();
 
-    const callbacks = this.getCallbacks('winner');
-    _.forEach(callbacks, callback => callback(socketMsg));
+    _.each(changes, change => {
+      const nominee = this.findNominee(change.nomination_id);
+      nominee.odds_numerator = change.odds_numerator;
+      nominee.odds_denominator = change.odds_denominator;
+      nominee.odds_expert = change.odds_expert;
+      nominee.odds_user = change.odds_user;
+    });
+
+    return this.packageUpResponse({msg: 'Success!'}, requestInfo);
+  }
+
+  private removeWinner(requestInfo: RequestInfo): void {
+    const winner_id = requestInfo.id;
+
+    const categoryForExistingWinner = this.existingCategoryForWinnerID(winner_id);
+    const winner = _.findWhere(categoryForExistingWinner.winners, {id: winner_id});
+    ArrayUtil.removeFromArray(categoryForExistingWinner.winners, winner);
+
+    const socketMsg = {
+      nomination_id: winner.nomination_id,
+      event_id: 1,
+      event_time: new Date(),
+      winner_id: winner.id
+    };
+
+    this.broadcastToChannel('remove_winner', socketMsg);
 
     this.sendUpdatedOdds();
   }
@@ -194,7 +290,7 @@ export class InMemoryDataService implements InMemoryDbService {
       event_id: 1,
       odds: odds_results,
     };
-    _.forEach(this.getCallbacks('odds'), callback => callback(odds_execution));
+    this.broadcastToChannel('odds', odds_execution);
   }
 
   private getVoters(): number[] {
@@ -202,102 +298,98 @@ export class InMemoryDataService implements InMemoryDbService {
     return _.uniq(fullList);
   }
 
-  private getCategoriesWithVotes(requestInfo: RequestInfo): Observable<any> {
-    return requestInfo.utils.createResponse$(() => {
-      console.log('HTTP GET override');
-
-      const dataEncapsulation = requestInfo.utils.getConfig().dataEncapsulation;
-
-      const entries = requestInfo.query.entries();
-      const person_id = entries.next().value[1][0];
-      const year = entries.next().value[1][0];
-
-      const data = [];
-
-      _.forEach(requestInfo.collection, category => {
-        const copyCategory: Category = {
-          id: category.id,
-          name: category.name,
-          points: category.points,
-          nominees: _.where(category.nominees, {year: +year}),
-          voted_on: this.getVoteForCategory(category.id, +person_id, +year),
-          winners: _.where(category.winners, {year: +year})
-        };
-        data.push(copyCategory);
-      });
-
-      const options: ResponseOptions = data ?
-        {
-          body: dataEncapsulation ? { data } : data,
-          status: STATUS.OK
-        } :
-        {
-          body: dataEncapsulation ? { } : data,
-          status: STATUS.OK
-        };
-      return InMemoryDataService.finishOptions(options, requestInfo);
-    });
+  // noinspection JSMethodCanBeStatic
+  private createNomineeCopy(nominee): any {
+    return {
+      id: nominee.id,
+      nominee: nominee.nominee,
+      context: nominee.context,
+      detail: nominee.detail,
+      category_id: nominee.category_id,
+      year: nominee.year,
+      odds_expert: nominee.odds_expert,
+      odds_user: nominee.odds_user,
+      odds_numerator: nominee.odds_numerator,
+      odds_denominator: nominee.odds_denominator,
+    };
   }
 
-  private getMaxYear(requestInfo: RequestInfo): Observable<any> {
-    return requestInfo.utils.createResponse$(() => {
-      console.log('HTTP GET override');
+  private getCategoriesWithVotes(requestInfo: RequestInfo): Observable<Response> {
+    console.log('HTTP GET override');
 
-      const dataEncapsulation = requestInfo.utils.getConfig().dataEncapsulation;
+    const person_id = requestInfo.query.get('person_id')[0];
+    const year = requestInfo.query.get('year')[0];
 
-      const maxYear = _.max(_.map(this.categories, category => _.max(_.map(category.nominees, nominee => nominee.year))));
-      const data = {
-        maxYear: maxYear
+    const data = [];
+
+    _.forEach(requestInfo.collection, category => {
+      const yearNum = +year;
+      const personIDNum = +person_id;
+      const copyCategory: Category = {
+        id: category.id,
+        name: category.name,
+        points: category.points,
+        nominees: _.chain(category.nominees)
+          .where({year: yearNum})
+          .map(this.createNomineeCopy)
+          .value(),
+        voted_on: this.getVoteForCategory(category.id, personIDNum, yearNum),
+        winners: _.where(category.winners, {year: yearNum})
       };
-
-      const options: ResponseOptions = data ?
-        {
-          body: dataEncapsulation ? { data } : data,
-          status: STATUS.OK
-        } :
-        {
-          body: dataEncapsulation ? { } : data,
-          status: STATUS.OK
-        };
-      return InMemoryDataService.finishOptions(options, requestInfo);
+      data.push(copyCategory);
     });
+
+    return this.packageUpResponse(data, requestInfo);
   }
 
-  private getEventsSinceDate(requestInfo: RequestInfo): Observable<Event[]> {
-    return requestInfo.utils.createResponse$(() => {
-      console.log('HTTP GET override');
+  private getMaxYear(requestInfo: RequestInfo): Observable<Response> {
+    console.log('HTTP GET override');
 
-      const dataEncapsulation = requestInfo.utils.getConfig().dataEncapsulation;
+    const maxYear = _.max(_.map(this.categories, category => _.max(_.map(category.nominees, nominee => nominee.year))));
+    const data = [{maxYear}];
 
-      const entries = requestInfo.query.entries();
-      const since_date = entries.next().value[1][0];
+    return this.packageUpResponse(data, requestInfo);
+  }
 
-      const data = requestInfo.collection;
-      // const data = _.filter(requestInfo.collection, event => event.event_time > since_date);
+  private getEventsSinceDate(requestInfo: RequestInfo): Observable<Response> {
+    console.log('HTTP GET override');
 
-      const options: ResponseOptions = data ?
-        {
-          body: dataEncapsulation ? { data } : data,
-          status: STATUS.OK
-        } :
-        {
-          body: dataEncapsulation ? { } : data,
-          status: STATUS.OK
-        };
-      return InMemoryDataService.finishOptions(options, requestInfo);
-    });
+    const entries = requestInfo.query.entries();
+    const since_date = entries.next().value[1][0];
+
+    const data = requestInfo.collection;
+    // const data = _.filter(requestInfo.collection, event => event.event_time > since_date);
+
+    return this.packageUpResponse(data, requestInfo);
   }
 
   private getVoteForCategory(category_id: number, person_id: number, year: number): number {
     const existingVote = _.findWhere(this.votes, {
-      category_id: category_id,
-      person_id: person_id,
-      year: year
+      category_id,
+      person_id,
+      year
     });
     return existingVote ? existingVote.nomination_id : undefined;
   }
 
-  private updateNomination(requestInfo: RequestInfo) {
+  genWinnerId(): number {
+    const winners = _.flatten(_.map(this.categories, c => c.winners));
+    if (winners.length === 0) {
+      return 1;
+    } else {
+      return _.max(_.map(winners, w => w.id)) + 1;
+    }
+  }
+
+  genVoteId(): number {
+    if (this.votes.length === 0) {
+      return 1;
+    } else {
+      return _.max(_.map(this.votes, w => w.id)) + 1;
+    }
+  }
+
+  private updateNomination(requestInfo: RequestInfo): Observable<Response> {
     const jsonBody = requestInfo.utils.getJsonBody(requestInfo.req);
     this.updateObject(jsonBody);
 
@@ -321,11 +413,19 @@ export class InMemoryDataService implements InMemoryDbService {
     });
   }
 
-  private existingCategoryForWinner(nomination_id: number): Category {
+  private existingCategoryForWinnerID(winner_id: number): Category {
     const results = _.filter(this.categories, category => {
-      return !!_.findWhere(category.winners, {nomination_id: nomination_id});
+      return !!_.findWhere(category.winners, {id: winner_id});
     });
     return _.first(results);
+  }
+
+  private findNominee(nomination_id: number): Nominee {
+    return _.chain(this.categories)
+      .map(category => category.nominees)
+      .flatten()
+      .findWhere({id: nomination_id})
+      .value();
   }
 
   private categoryForNominee(nomination_id: number): Category {
@@ -335,30 +435,32 @@ export class InMemoryDataService implements InMemoryDbService {
     return _.first(results);
   }
 
-  private updateVote(requestInfo: RequestInfo, existingVote: Vote) {
-    const jsonBody = requestInfo.utils.getJsonBody(requestInfo.req);
-    existingVote.nomination_id = jsonBody.nomination_id;
-
-    const options: ResponseOptions = {
-      body: existingVote,
-      status: STATUS.OK
-    };
-
-    const finishedOptions = InMemoryDataService.finishOptions(options, requestInfo);
-
-    return requestInfo.utils.createResponse$(() => finishedOptions);
-  }
-
-  private updateObject(jsonBody: any) {
+  private updateObject(jsonBody: any): void {
     const id = jsonBody.id;
     const odds_expert = jsonBody.odds_expert;
 
-    _.forEach(this.categories, function(category) {
-      _.forEach(category.nominees, function(nominee) {
+    _.forEach(this.categories, category => {
+      _.forEach(category.nominees, nominee => {
         if (nominee.id === id) {
           nominee.odds_expert = odds_expert;
         }
       });
     });
   }
+
+  // noinspection JSMethodCanBeStatic
+  private getBody(requestInfo): any {
+    return requestInfo.utils.getJsonBody(requestInfo.req);
+  }
+
+  private packageUpResponse(body, requestInfo): Observable<Response> {
+    const options: ResponseOptions = {
+      body,
+      status: STATUS.OK
+    };
+    const finishedOptions = InMemoryDataService.finishOptions(options, requestInfo);
+
+    return requestInfo.utils.createResponse$(() => finishedOptions);
+  }
+
 }
