@@ -1,5 +1,4 @@
-import { Injectable } from '@angular/core';
-import {_} from 'underscore';
+import {Injectable} from '@angular/core';
 import {InMemoryDataService} from './in-memory-data-service';
 
 @Injectable()
@@ -8,15 +7,23 @@ export class SocketServiceMock {
   constructor(private inMemoryDB: InMemoryDataService) {
   }
 
-  on(channel, callback) {
+  init(): void {
+  }
+
+  on(channel, callback): void {
     this.inMemoryDB.on(channel, callback);
   }
 
-  removeListener(channel, callback) {
+  removeListener(channel, callback): void {
     this.inMemoryDB.removeCallback(channel, callback);
   }
 
-  isConnected() {
+  emit(channel, msg): void {
+    console.log('Message sent to channel ' + channel + ': ' + JSON.stringify(msg));
+    this.inMemoryDB.broadcastToChannel(channel, msg);
+  }
+
+  isConnected(): boolean {
     return true;
   }
 }
